moves query constants to types
